requirements.txt
Flask
fake-useragent
bs4
requests
pillow
selenium
psycopg2-binary